Big-ass refactoring for v0.5

+++ authserver/src/internal/dtos/auth_context.go
@@ -29,7 +29,7 @@ type AuthContext struct {
 	AcrLevel            string
 	AuthMethods         string
 	AuthTime            time.Time
-	UserId              uint
+	UserId              int64
 	AuthCompleted       bool
 }
